pyproject.toml
[tool.poetry]
name = "cherrycandy"
version = "0.0.1"
description = "Discord bot"
authors = ["kiki7000"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.0"
termcolor = "^1.1.0"
dico-api = { git = "https://github.com/dico-api/dico", branch = "master" }
dico-interaction = { git = "https://github.com/dico-api/dico-interaction", branch = "master" }
dico-dp = "^0.0.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pylint = "^2.10.2"
pycodestyle = "^2.7.0"
black = {version = "^21.8b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.4"]
build-backend = "poetry.core.masonry.api"
